+++ packages/client/ui/react-ui/src/components/auth/methods/email/EmailAuthFlow.tsx
@@ -1,15 +1,18 @@
+import { useState } from "react";
 import { EmailOTPInput } from "./EmailOTPInput";
 import { EmailSignIn } from "./EmailSignIn";
 import { useAuthForm } from "@/providers/auth/AuthFormProvider";
+import type { OtpEmailPayload } from "@/types/auth";
 
 export function EmailAuthFlow() {
     const { step } = useAuthForm();
+    const [otpEmailData, setOtpEmailData] = useState<OtpEmailPayload | null>(null);
 
     if (step === "otp") {
-        return <EmailOTPInput />;
+        return <EmailOTPInput otpEmailData={otpEmailData} setOtpEmailData={setOtpEmailData} />;
     }
     if (step === "initial") {
-        return <EmailSignIn />;
+        return <EmailSignIn setOtpEmailData={setOtpEmailData} />;
     }
 
     return null;
