+++ packages/client/ui/react-ui/src/components/index.ts
@@ -4,4 +4,4 @@ export * from "./CrossmintNFTDetail";
 export * from "./embed";
 export * from "./hosted";
 
-export { AuthForm } from "./auth/AuthForm";
+export { AuthForm as EmbeddedAuthForm } from "./auth/AuthForm";
