+++ packages/client/ui/react-ui/src/providers/auth/AuthFormProvider.tsx
@@ -1,5 +1,4 @@
 import { createContext, useContext, useState, type ReactNode } from "react";
-import type { OtpEmailPayload } from "@/types/auth";
 import type { AuthMaterial } from "@crossmint/common-sdk-auth";
 import type { UIConfig } from "@crossmint/common-sdk-base";
 import type { LoginMethod } from "../CrossmintAuthProvider";
@@ -8,18 +7,13 @@ type AuthStep = "initial" | "walletMethod" | "otp" | "qrCode";
 
 interface AuthFormContextType {
     step: AuthStep;
-    email: string;
     apiKey: string;
     baseUrl: string;
     fetchAuthMaterial: (refreshToken: string) => Promise<AuthMaterial>;
     appearance?: UIConfig;
     loginMethods: LoginMethod[];
-    otpEmailData: OtpEmailPayload | null;
     setStep: (step: AuthStep) => void;
-    setEmail: (email: string) => void;
     setDialogOpen: (open: boolean) => void;
-    setOtpEmailData: (data: OtpEmailPayload | null) => void;
-    resetState: () => void;
 }
 
 type ContextInitialStateProps = {
@@ -45,29 +39,17 @@ export const AuthFormProvider = ({
     children,
     initialState,
 }: { children: ReactNode; initialState: ContextInitialStateProps }) => {
-    const [otpEmailData, setOtpEmailData] = useState<OtpEmailPayload | null>(null);
     const [step, setStep] = useState<AuthStep>("initial");
-    const [email, setEmail] = useState("");
-
-    const resetState = () => {
-        setStep("initial");
-        setEmail("");
-    };
 
     const value: AuthFormContextType = {
         step,
-        email,
         apiKey: initialState.apiKey,
         baseUrl: initialState.baseUrl,
         fetchAuthMaterial: initialState.fetchAuthMaterial,
         appearance: initialState.appearance,
-        otpEmailData,
         loginMethods: initialState.loginMethods,
         setDialogOpen: initialState.setDialogOpen ?? (() => {}),
         setStep,
-        setEmail,
-        setOtpEmailData,
-        resetState,
     };
 
     return <AuthFormContext.Provider value={value}>{children}</AuthFormContext.Provider>;
