+++ packages/client/ui/react-ui/src/types/auth.ts
@@ -1,11 +1,3 @@
-export type OAuthProvider = "google" | "farcaster" | "discord";
-
-export interface OAuthUrlParams {
-    provider: OAuthProvider;
-    apiKey: string;
-    baseUrl: string;
-}
-
 export type OtpEmailPayload = {
     email: string;
     state: string;
