+++ packages/common/auth/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./constants";
+export * from "./isEmailValid";
